+++ elmyra.ip.access.epo/CHANGES.rst
@@ -10,6 +10,7 @@ development
 - Properly handle errors when IFI Claims delivers empty document for bibliographic data, e.g. IN268402A
 - Make IFI Claims data model grok patent citations
 - Enable documents from office “KR” for IFI Claims bibliographic data interface
+- Improve language priorization for fulltexts from IFI Claims: EN, DE, others
 
 2017-03-28 0.132.3
 ==================